Reorganize class structure and fix HabitEntry access modifier to be public
