+++ HabitTracker.tristenCodes/DTO/HabitEntry.cs
@@ -1,30 +1,8 @@
-class HabitEntry
+namespace DTO;
+
+public class HabitEntry
 {
-    private string _habit = "";
-    private DateTime _date;
+    public string Habit { get; set; } = "";
+    public DateTime Date { get; set; }
 
-    public string Habit
-    {
-        get
-        {
-            return _habit;
-        }
-        set
-        {
-            _habit = value;
-        }
-    }
-
-    public DateTime Date
-    {
-        get
-        {
-            return _date;
-        }
-        set
-        {
-
-            DateTime DT = DateTime.ParseExact(value, "dd-MM-yyyy", CultureInfo.InvariantCulture);
-        }
-    }
 }
